+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/transformer/SumAggregator.scala
@@ -1,5 +1,6 @@
 package com.twitter.finatra.streams.transformer
 
+import com.twitter.conversions.DurationOps._
 import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finatra.streams.transformer.FinatraTransformer.WindowStartTime
 import com.twitter.finatra.streams.transformer.domain._
@@ -18,7 +19,7 @@ class SumAggregator[K, V](
   windowSize: Duration,
   allowedLateness: Duration,
   queryableAfterClose: Duration,
-  windowStart: (Time, K, V) => Long,
+  windowStart: (Time, K, V) => Time,
   countToAggregate: (K, V) => Int,
   emitOnClose: Boolean = true,
   maxActionsPerTimer: Int = 25000)
@@ -33,10 +34,6 @@ class SumAggregator[K, V](
     with StateStoreImplicits
     with IteratorImplicits {
 
-  private val windowSizeMillis = windowSize.inMillis
-  private val allowedLatenessMillis = allowedLateness.inMillis
-  private val queryableAfterCloseMillis = queryableAfterClose.inMillis
-
   private val restatementsCounter = statsReceiver.counter("numRestatements")
   private val closedCounter = statsReceiver.counter("closedWindows")
   private val expiredCounter = statsReceiver.counter("expiredWindows")
@@ -50,26 +47,26 @@ class SumAggregator[K, V](
 
   override def onMessage(time: Time, key: K, value: V): Unit = {
     val windowedKey = TimeWindowed.forSize(
-      startMs = windowStart(time, key, value),
-      sizeMs = windowSizeMillis,
+      start = windowStart(time, key, value),
+      size = windowSize,
       value = key
     )
 
     val count = countToAggregate(key, value)
-    if (windowedKey.isLate(allowedLatenessMillis, Watermark(watermark.timeMillis))) {
+    if (windowedKey.isLate(allowedLateness, Watermark(watermark.timeMillis))) {
       restatementsCounter.incr()
       forward(windowedKey, WindowedValue(Restatement, count))
     } else {
       val newCount = stateStore.increment(windowedKey, count)
       if (newCount == count) {
-        val closeTime = windowedKey.startMs + windowSizeMillis + allowedLatenessMillis
+        val closeTime = windowedKey.start + windowSize + allowedLateness
         if (emitOnClose) {
-          timerStore.addTimer(Time(closeTime), Close, windowedKey.startMs)
+          timerStore.addTimer(closeTime, Close, windowedKey.start.millis)
         }
         timerStore.addTimer(
-          Time(closeTime + queryableAfterCloseMillis),
+          closeTime + queryableAfterClose,
           Expire,
-          windowedKey.startMs
+          windowedKey.start.millis
         )
       }
     }
@@ -80,14 +77,15 @@ class SumAggregator[K, V](
     timerMetadata: TimerMetadata,
     windowStartMs: WindowStartTime
   ): Unit = {
+    val windowStart = Time(windowStartMs)
     val hourlyWindowIterator = stateStore.range(
-      TimeWindowed.forSize(windowStartMs, windowSizeMillis, keyRangeStart),
-      TimeWindowed.forSize(windowStartMs + 1, windowSizeMillis, keyRangeStart)
+      TimeWindowed.forSize(windowStart, windowSize, keyRangeStart),
+      TimeWindowed.forSize(windowStart + 1.millis, windowSize, keyRangeStart)
     )
 
     try {
       if (timerMetadata == Close) {
-        onClosed(windowStartMs, hourlyWindowIterator)
+        onClosed(windowStart, hourlyWindowIterator)
       } else {
         onExpired(hourlyWindowIterator)
       }
@@ -97,7 +95,7 @@ class SumAggregator[K, V](
   }
 
   private def onClosed(
-    windowStartMs: Long,
+    windowStart: Time,
     windowIterator: KeyValueIterator[TimeWindowed[K], Int]
   ): Unit = {
     windowIterator