+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/transformer/internal/domain/Timer.scala
@@ -1,14 +1,14 @@
 package com.twitter.finatra.streams.transformer.internal.domain
 
 import com.twitter.finatra.streams.converters.time._
-import com.twitter.finatra.streams.transformer.domain.TimerMetadata
+import com.twitter.finatra.streams.transformer.domain.{Time, TimerMetadata}
 
 /**
  * @param time Time to fire the timer
  */
-case class Timer[K](time: Long, metadata: TimerMetadata, key: K) {
+case class Timer[K](time: Time, metadata: TimerMetadata, key: K) {
 
   override def toString: String = {
-    s"Timer(${metadata.getClass.getName} $key @${time.iso8601Millis})"
+    s"Timer(${metadata.getClass.getName} $key @${time.millis.iso8601Millis})"
   }
 }