+++ kafka-streams/kafka-streams/src/test/scala/com/twitter/unittests/integration/compositesum/UserClicksTopologyFeatureTest.scala
@@ -2,19 +2,8 @@ package com.twitter.unittests.integration.compositesum
 
 import com.twitter.conversions.DurationOps._
 import com.twitter.finatra.streams.tests.{FinatraTopologyTester, TopologyFeatureTest}
-import com.twitter.finatra.streams.transformer.domain.{
-  FixedTimeWindowedSerde,
-  TimeWindowed,
-  WindowClosed,
-  WindowOpen,
-  WindowedValue,
-  WindowedValueSerde
-}
-import com.twitter.unittests.integration.compositesum.UserClicksTypes.{
-  ClickTypeSerde,
-  NumClicksSerde,
-  UserIdSerde
-}
+import com.twitter.finatra.streams.transformer.domain._
+import com.twitter.unittests.integration.compositesum.UserClicksTypes.{ClickTypeSerde, NumClicksSerde, UserIdSerde}
 import org.joda.time.DateTime
 
 class UserClicksTopologyFeatureTest extends TopologyFeatureTest {
@@ -35,8 +24,8 @@ class UserClicksTopologyFeatureTest extends TopologyFeatureTest {
 
   test("windowed clicks") {
     val userId1 = 1
-    val firstHourStartMillis = new DateTime("2018-01-01T00:00:00Z").getMillis
-    val fifthHourStartMillis = new DateTime("2018-01-01T05:00:00Z").getMillis
+    val firstHourStartTime = Time.create(new DateTime("2018-01-01T00:00:00Z"))
+    val fifthHourStartTime = Time.create(new DateTime("2018-01-01T05:00:00Z"))
 
     userIdToClicksTopic.pipeInput(userId1, 100)
     userIdToClicksTopic.pipeInput(userId1, 200)
@@ -47,15 +36,15 @@ class UserClicksTopologyFeatureTest extends TopologyFeatureTest {
 
     topologyTester.advanceWallClockTime(30.seconds)
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 100)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 100)),
       WindowedValue(WindowOpen, 1))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 300)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 300)),
       WindowedValue(WindowOpen, 3))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 200)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 200)),
       WindowedValue(WindowOpen, 2))
 
     userIdToClicksTopic.pipeInput(userId1, 100)
@@ -64,34 +53,34 @@ class UserClicksTopologyFeatureTest extends TopologyFeatureTest {
 
     topologyTester.advanceWallClockTime(5.hours)
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 100)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 100)),
       WindowedValue(WindowOpen, 2))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 300)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 300)),
       WindowedValue(WindowOpen, 4))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 200)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 200)),
       WindowedValue(WindowOpen, 3))
 
     userIdToClicksTopic.pipeInput(userId1, 1)
     topologyTester.advanceWallClockTime(30.seconds)
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(fifthHourStartMillis, UserClicks(userId1, clickType = 1)),
+      TimeWindowed.hourly(fifthHourStartTime, UserClicks(userId1, clickType = 1)),
       WindowedValue(WindowOpen, 1))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 100)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 100)),
       WindowedValue(WindowClosed, 2))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 200)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 200)),
       WindowedValue(WindowClosed, 3))
 
     hourlyWordAndCountTopic.assertOutput(
-      TimeWindowed.hourly(firstHourStartMillis, UserClicks(userId1, clickType = 300)),
+      TimeWindowed.hourly(firstHourStartTime, UserClicks(userId1, clickType = 300)),
       WindowedValue(WindowClosed, 4))
   }
 }